backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
python-dotenv==1.2.2
google-genai>=1.0.0
slowapi==0.1.9
